pyproject.toml
[project]
name = "longbo-cloud-autobot"
version = "0.1.0"
description = "Autonomous WordPress autopublisher for longbo.cloud"
authors = [
    {name = "AI"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "typer[all] (>=0.17.4,<0.18.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "backoff (>=2.2.1,<3.0.0)",
    "feedparser (>=6.0.12,<7.0.0)",
    "selectolax (>=0.3.34,<0.4.0)",
    "lxml (>=6.0.1,<7.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "sqlmodel (>=0.0.25,<0.0.26)",
    "simhash (>=2.1.2,<3.0.0)",
    "rapidfuzz (>=3.14.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "orjson (>=3.11.3,<4.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "yarl (>=1.20.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[project.scripts]
longbo = "autobot.cli:main"

[tool.poetry]
packages = [{ include = "autobot" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
